First pass at the dependency algorithm
Will need to setup some tests to verify this. Maybe break down the 
asteroids game into a set of modules: Game field (with location, mobile, 
and the systems to update just those), Asteroids (colliders, asteroids, 
and maybe explosions), Player (player ship and bullet components, along 
with the code to handle score and life).